Clean up of VS2015 warnings - Part 1
Fix C6283: <variable> is allocated with array new [], but deleted with scalar delete
Fix C4244: conversion from 'int' to 'size_t' ', possible loss of data for length variables
Remove redundant  #include <olectlid.h>

+++ native/ax/FileSource/NetStreamData.cpp
@@ -132,7 +132,7 @@ HRESULT CNetStreamData::OpenConnection( )
 	DbgLog((LOG_TRACE, 2, TEXT("connect succeeded")));
 	char data[512];
 	// Check if we need to use the Unicode version
-	int wlen = wcslen(m_pFileName);
+	size_t wlen = wcslen(m_pFileName);
 	int needUni = 0;
 	for (int i = 0; i < wlen; i++)
 	{
@@ -147,7 +147,7 @@ HRESULT CNetStreamData::OpenConnection( )
 		strcpy(data, "OPENW ");
 	else
 		strcpy(data, "OPEN ");
-	int dataIdx = strlen(data);
+	size_t dataIdx = strlen(data);
 	for (int i = 0; i < wlen; i++, dataIdx+=(needUni?2:1))
 	{
 		if (needUni)
@@ -189,7 +189,7 @@ void CNetStreamData::CloseConnection()
 	if (m_sd != INVALID_SOCKET)
 	{
 		char* data = "QUIT\r\n";
-		int dataSize = strlen(data);
+		size_t dataSize = strlen(data);
 		send(m_sd, data, dataSize, 0);
 		closesocket(m_sd);
 		m_sd = INVALID_SOCKET;
@@ -286,7 +286,7 @@ HRESULT CNetStreamData::SetPointer(LONGLONG llPos)
 	}
 	char data[512];
 	sprintf(data, "SIZE\r\n");
-	int dataSize = strlen(data);
+	size_t dataSize = strlen(data);
 	if (send(m_sd, data, dataSize, 0) < dataSize)
 	{
 		DbgLog((LOG_TRACE, 2, TEXT("socket write failed, reopening...")));
@@ -370,7 +370,7 @@ HRESULT CNetStreamData::Read(PBYTE pbBuffer, DWORD dwBytesToRead, BOOL bAlign, L
 		temp[--pos] = (TCHAR) digit+L'0';
 	} while (li.QuadPart);
 	sprintf(data, "READ %s %d\r\n", temp+pos, dwBytesToRead);
-	int dataSize = strlen(data);
+	size_t dataSize = strlen(data);
 	if (send(m_sd, data, dataSize, 0) < dataSize)
 	{
 		DbgLog((LOG_TRACE, 2, TEXT("socket write failed, reopening connection...")));
@@ -442,7 +442,7 @@ LONGLONG CNetStreamData::Size( LONGLONG *pSizeAvailable , LONGLONG *pOverwritten
 	LONGLONG SizeAvailable;
 	char data[512];
 	sprintf(data, "SIZE\r\n");
-	int dataSize = strlen(data);
+	size_t dataSize = strlen(data);
 	if (send(m_sd, data, dataSize, 0) < dataSize)
 	{
 		DbgLog((LOG_TRACE, 3, TEXT("socket write failed, reopening...")));
