+++ native/ax/StrmDeMux/SDeMuxOutPin.cpp
@@ -695,7 +695,7 @@ void CSDeMuxOutPin::open_dump( )
 {
 	char fname[MAX_PATH];
 	char path[MAX_PATH]={0};
-	int len;
+	size_t len;
 	sprintf( fname, "DUMP_DATA.ENABLE" );
 	fd = fopen( fname, "r" );
 	if ( fd == NULL ) return;
@@ -712,7 +712,7 @@ void CSDeMuxOutPin::open_dump( int index )
 {
 	char fname[MAX_PATH];
 	char path[MAX_PATH]={0};
-	int len;
+	size_t len;
 	sprintf( fname, "DUMP_DATA.ENABLE" );
 	fd = fopen( fname, "r" );
 	if ( fd == NULL ) return;
@@ -757,7 +757,7 @@ HRESULT CSDeMuxOutPin::FillBuffer(IMediaSample *pSample)
 
 	static FILE* fp=NULL;
 	static int frame = 0;
-	int ct;
+	size_t ct;
 	if ( frame == 0 )
 		fp = fopen( "video.bin", "rb" );
 	if ( fp != NULL )
