+++ native/exe/SageLauncher/launcher.cpp
@@ -143,8 +143,8 @@ static TCHAR toHex(int nibble)
 void writeEscapedPropString(FILE* fp, const LPWSTR wstr)
 {
 	static const TCHAR specialSaveChars[] = "=: \t\r\n\f#!";
-    int len = wcslen(wstr);
-	int x;
+	size_t len = wcslen(wstr);
+	size_t x;
     for(x=0; x<len; x++) 
 	{
 		if (wstr[x] == ' ')
@@ -256,7 +256,7 @@ BOOL CheckForMutex(HANDLE *pMutex, LPCSTR szMutexName, LPSTR appName)
 
 void removeTrailingWS(char* line)
 {
-	int len = strlen(line);
+	size_t len = strlen(line);
 	while (len > 0 && (line[len - 1] == ' ' || line[len - 1] == '\r' || line[len-1] == '\n'))
 	{
 		line[len - 1] = '\0';
@@ -625,8 +625,8 @@ int WINAPI WinMain( HINSTANCE hInst, 	/*Win32 entry-point routine */
 	// Fix the current working directory to be where the EXE is located
 	LPTSTR folderPath = new TCHAR[2048];
 	GetModuleFileName(NULL, folderPath, 2048);
-	int appLen = strlen(folderPath);
-	for (int i = appLen - 1; i > 0; i--)
+	size_t appLen = strlen(folderPath);
+	for (size_t i = appLen - 1; i > 0; i--)
 	{
 		if (folderPath[i] == '\\')
 		{
@@ -919,8 +919,8 @@ int launchJVMSage(LPSTR lpszCmdLine, HWND hWnd, BOOL bClient, BOOL bService)
 	
 	LPTSTR appPath = new TCHAR[2048];
 	GetModuleFileName(NULL, appPath, 2048);
-	int appLen = strlen(appPath);
-	for (int i = appLen - 1; i > 0; i--)
+	size_t appLen = strlen(appPath);
+	for (size_t i = appLen - 1; i > 0; i--)
 	{
 		if (appPath[i] == '\\')
 		{
